♻️ [Refactor] map관련 뷰, 뷰모델 분리
